requirements.txt
# Core dependencies for ICENews ingestion
scrapfly-sdk>=0.8.5
jmespath>=1.0.1

# Optional but recommended (used by scraper)
loguru>=0.7.0

# Web app
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
httpx>=0.25.0
